Temporary hack to prevent consumers going bankrupt
At our current development stage production chains are not closed loops, which means that "consumers" at the end of the chain are buying but not selling and, as a result, going bankrupt almost immediately. Once we implement bankruptcy (coming imminently), this would very rapidly collapse the entire market. The temporary solution is to hardcode the amount of money they have.