Fix issue where production cost constantly pushed up price
Previously production cost was being simply added in to order prices. While the intent was to have price beliefs for produced goods actually function as a belief in how much profit an Agent should be making, this was not consistently applied when updating price beliefs, and in fact would have been very difficult to do so properly. To work around this, cost now functions as a floor on order prices, where regardless of current price beliefs no order will be priced below cost+10% -- a completely arbitrary minimum profit margin.