MarketHistory object manages days itself
The MarketHistory object now expects that calling code will "open" a day of trading, add trade summaries for individual goods for the day, and then "close" the day to actually commit it to the history. Currently there is no checking to ensure that a day actually has been closed before the next one is opened, which would silently discard the entire day of trading.