Agents now try to buy/sell at low/high prices
Agents examine the Market history for each good they wish to trade. If the most current mean price is high (relative to the 30-day high/low prices), they are more likely to sell but less likely to buy; if price is low, the opposite is true. This is reflected in the quantities in the orders they place.

There is currently an issue where all trading ceases entirely if the mean price drops too close to the low, as this drops the "ratio" close enough to 0 that Agents only want to sell 0 of the good; this probably also impacts buyers if the mean price climbs too close to the high, though this has not yet been observed.