Add Job limit
This parameter limits the maximum number of runs of a Job a given Agent can perform in a single round. In addition to the new limit property, we now have a generator on Job that will only run for the maximum number of runs -- or forever if there is no limit.