Handle divide-by-zero error in history
In the case of the high and low of a good being the same -- for example, on the first day it was traded if only a single unit was bought -- we wound up trying to divide by zero. That obviously does not work. Now we hard-code the result to say that the mean price is at the mid-point (i.e. a ratio of 0.5).